Dissect/Decode Content based on Name provided by app/user within Data
As opposed to overriding the PacketListField with a custom getfield to give the
underlayer of Name to it. Sewing layers post_dissect(ion) is painful.
SignatureValue guess is also moved within guessing the Data
PacketListField as SignatureInfo is expected to be available beore
SignatureValue. This should allow users to manipulate Content value
after dissection.